Add x64 support.

+++ GMOD/panelwrapper.h
@@ -1,8 +1,8 @@
 class VPanelWrapper
 {
 public:
-	const char* GetName(unsigned int VPanel) {
-		typedef const char*(__thiscall* fn)(void*, unsigned int); // фиксим варнинг
+	const char* GetName(uintptr_t VPanel) {
+		typedef const char*(__thiscall* fn)(void*, uintptr_t);
 		return VMT.getvfunc<fn>(this, 36)(this, VPanel);
 	}
